Upgrade to Flask service.
This code creates a simple Flask app with two routes: / and /generate. The / route displays a form where the user can enter a link and submit it to generate a QR code. The /generate route takes the submitted link and uses the qrcode library to generate a QR code, which is then returned as an image.

To run this app on your system, make sure you have Flask and the qrcode library installed by running the commands pip install Flask and pip install qrcode. Then, save the above code to a file named app.py and run it using the command python app.py. The app should now be running on your Raspberry Pi and accessible at http://serveraddress:5000.