add decoding
This code adds a new route /decode to the Flask app that allows users to upload a QR code image. The uploaded image is then decoded using the pyzbar library and the decoded text is returned as the response.

To use this updated version of the app, make sure you have the pyzbar library installed by running the command pip install pyzbar. Then, save the above code to a file named app.py and run it using the command python app.py. The updated app should now be running on your system and accessible at http://<serveraddress>:5000.