add logo function
This code modifies the /generate route to add a logo to the center of the generated QR code. The logo.png file should be an image of the logo you want to add and should be located in the same directory as the app.py file.

After making these changes and saving the updated code to app.py, you can run the app using the command python app.py. The updated app should now generate QR codes with a logo in the center.