add css
This code adds some basic CSS styles to the page to change the font, add a background color to the heading, and add some margins to the content.

You can further customize the appearance of the page by adding more CSS styles or by using a CSS framework like Bootstrap to quickly create a visually appealing layout.